scsi:ufs:remove sanity check
UFS descriptors assumed to have a 2 bytes header:
 - Offset 0 - descriptor size
 - Offset 1 - descriptor id
However, the ufs spec does not mandates this, and as a matter of
fact, QUERY_DESC_DESC_TYPE_OFFSET is not being used anymore now.
And future descriptors no longer follow this assumption. So
remove this as preparation of introducing the FBO descriptor.

Signed-off-by: lijiaming3 <lijiaming3@xiaomi.com>
Reviewed-by: Avri Altman <avri.altman@wdc.com>